:SubstituteChoices: ENH: Include count of match replacements in current line in query
As there's no special highlighting of the current match, and previous replacements are not reflected in the current substitution, the user has to count in his mind which match this is about, which is very error prone.
Move the accounting of s:lnum from s:ShowContext() to s:Replace() and add s:matchesInLineCnt.