CHG: Escaping of whitespace and backslashes in :SubstituteMultiple is not necessary any longer
(But now the full /{pattern}/{string}/ needs to be given; you cannot omit e.g. the trailing /). This parsing better matches the user expectation. :SubstituteWildcard still requires escaping, though.
Use <q-args> instead of <f-args> for the command definition, take a single a:arguments string in PatternsOnText#Pairs#SubstituteMultiple(). Instead of map()-ing each split argument through ingo#cmdargs#substitute#Parse(), iteratively parse the head /{pattern}/{string}/ pair (with a custom definition of flagsExpr that matches the following one(s)).
One complication is the trailing flags / count; adding that to flagsExpr would add ambiguities (# is both a flag and valid separator), and it's hard to tighten the /.../.../ pattern because the previous stuff already uses capture groups, so we cannot use a backreference easily. Instead, reuse the existing parser in s:ParseArguments(), and feed it a simple whitespace-split (for flags and count, this isn't critical). In order to keep the original whitespace characters, don't re-join the split, but instead use the parsed flags and count to remove those from the original arguments.