Tweak rotate and shift expressions to work as expected with memoization
So far I had assumed that the contents of s:uniqueMatches are identical to the first elements in context.matches, but if the matches are scrambled (i.e. with repeating matches already in the first few positions), this is not the case.
We actually need to fully index into s:uniqueMatches for rotation, bypassing context.matches altogether.
For shifting, perform the bounds check to s:uniqueMatches, but then apply the offset to the first such found match.