s:ConfirmQuery(): Abort when an invalid accelerator has been input and query again
It's too easy to mistype an accelerator; having all remaining matches kept as-is is unfriendly, as subsitutions from the middle of a line are problematic to restart and undoing and redoing all previous choices may mean a lot of lost work.
Instead, indicate the invalid input via a beep and query again. The easiest implementation is via tail-recursive call (in the hope that not too many wrong sequential inputs happen and smash the stack).