s:ShowContext(): Show the [part of the] current line with the current match highlighted with IncSearch
And other matches with Search, for a better overview of where we are. Relying on 'cursorline' / 'cursorcolumn' has been a poor kludge, and the fallback of :print'ing the line also isn't good (esp. with overly long lines).
We actually have a nice implementation of highlighting search matches in the current line for SearchFoldsWithoutOpening.vim; reuse that now that all parts have been moved into ingo-library. With this, the output never exceeds one statusline (which still should provide enough context), and the current match is highlighted differently (IncSearch vs. Search).